+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 This file keeps track of changes between releases for the bankrupt project
 which adheres to [semantic versioning](https://semver.org).
 
+## unreleased
+
+* Add upload support for PDF files.
+
 ## v2.0.1 2019-07-30
 
 Add support for uploading font files (`eot`, `ttf`, `woff`, and `woff2`) to